remove ErrMiddleware
conditional middlewares were a good idea in paper, however they need more
than just an error to work, the WriteHeader call needs to be deferred
until the last middleware is ran.

this can be achieved with a wrapper to the writer object, but the whole
purpose of the ErrHandler was avoiding that very thing.

reverting back to simple handlers was easier than expected